♻️ Common store: Split control-actions to new file
