+++ modules/ppcp-settings/resources/js/data/common/index.js
@@ -5,6 +5,7 @@ import { STORE_NAME } from './constants';
 import reducer from './reducer';
 import * as selectors from './selectors';
 import * as actions from './actions';
+import * as thunkActions from './actions-thunk';
 import * as hooks from './hooks';
 import { resolvers } from './resolvers';
 import { controls } from './controls';
@@ -19,7 +20,7 @@ export const initStore = () => {
 	const store = createReduxStore( STORE_NAME, {
 		reducer,
 		controls: { ...wpControls, ...controls },
-		actions,
+		actions: { ...actions, ...thunkActions },
 		selectors,
 		resolvers,
 	} );
